Allow calling .autodiff() multiple times
This only works if the expression to differentiate has been
re-evaluated. If you call .autodiff() more than once in a row without
re-evaluating the expression, you will not get the desired results.

Also, there seems to be something wrong with using the result of an
expression multiple times in another expression and differentiating with
respect to this new expression. Gotta investigate this.